refactor: change image Ivan
