+++ src/components/Layouts/MainLayout/MatchingFeed/IsMatchedModal/IsMatchedModal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import avatar from '../../../../../assets/user/avatar Ivan.svg';
+import avatar from '../../../../../assets/user/avatarIvan.svg';
 import styles from './IsMatchedModal.module.css';
 import { ModalMatchedProps } from '../interfaces';
